backend/requirements.txt
boto3==1.34.34
botocore==1.34.34
pydantic==2.12.5
langdetect==1.0.9
bleach==6.1.0
python-dateutil==2.8.2
hypothesis==6.98.0
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
httpx==0.28.1
python-dotenv==1.0.0
